+++ apps/next-client/src/app/[locale]/stripe/@checkout/components/index.ts
