+++ apps/next-client/src/app/[locale]/stripe/@checkout/success/page.tsx
