+++ apps/next-client/src/app/[locale]/stripe/page.tsx
@@ -1,15 +1,16 @@
 import { Fragment } from 'react';
+import type { Metadata } from 'next';
 import { PageHeading } from '@csl/shared-fe';
-import { StripePageLinks } from '@/constants';
-import { PageLink } from '@/components';
+
+export const metadata: Metadata = {
+  title: 'Stripe',
+  description: 'Stripe payment integration example'
+};
 
 export default function StripePage() {
   return (
     <Fragment>
       <PageHeading title="Stripe" />
-      {StripePageLinks.map((link, idx) => (
-        <PageLink text={link.title} href={link.href} key={idx} />
-      ))}
     </Fragment>
   );
 }
